splitting up ASTModels from DAGModels
