+++ src/Soss.jl
@@ -53,8 +53,8 @@ include("noted.jl")
 include("core/models/abstractmodel.jl")
 include("core/statement.jl")
 include("core/models/astmodel/astmodel.jl")
+include("core/models/dagmodel/dagmodel.jl")
 include("core/models/model.jl")
-# include("core/models/jointdistribution.jl")
 include("core/canonical.jl")
 include("core/dependencies.jl")
 include("core/toposort.jl")
