+++ src/core/models/conditional.jl
@@ -1,5 +1,5 @@
 struct ConditionalModel{A,B,M,Argvals,Obs} <: AbstractModel{A,B,M,Argvals,Obs}
-    model :: DAGModel{A,B,M}
+    model :: AbstractModel{A,B,M, Nothing, Nothing}
     argvals :: Argvals
     obs :: Obs
 end
@@ -22,20 +22,20 @@ function observed(cm::ConditionalModel{A,B,M,Argvals,Obs}) where {A,B,M,Argvals,
     keys(schema(Obs))
 end
 
-DAGModel(c::ConditionalModel) = c.model
+Model(c::ConditionalModel) = c.model
 
-ConditionalModel(m::DAGModel) = ConditionalModel(m,NamedTuple(), NamedTuple())
+ConditionalModel(m::AbstractModel) = ConditionalModel(m,NamedTuple(), NamedTuple())
 
-(m::DAGModel)(nt::NamedTuple) = ConditionalModel(m)(nt)
+(m::AbstractModel)(nt::NamedTuple) = ConditionalModel(m)(nt)
 
 (cm::ConditionalModel)(nt::NamedTuple) = ConditionalModel(cm.model, merge(cm.argvals, nt), cm.obs)
 
-(m::DAGModel)(;argvals...)= m((;argvals...))
+(m::AbstractModel)(;argvals...)= m((;argvals...))
 
-(m::DAGModel)(args...) = m(NamedTuple{Tuple(m.args)}(args...))
+(m::AbstractModel)(args...) = m(NamedTuple{Tuple(m.args)}(args...))
 
 import Base
 
-Base.:|(m::DAGModel, nt::NamedTuple) = ConditionalModel(m) | nt
+Base.:|(m::AbstractModel, nt::NamedTuple) = ConditionalModel(m) | nt
 
 Base.:|(cm::ConditionalModel, nt::NamedTuple) = ConditionalModel(cm.model, cm.argvals, merge(cm.obs, nt))
